Nuova UI sezione settings
- Ridefinita la classe Settings come JDialog
- Divise in classi le sezioni dei settings
- Migliorie generali alla sintassi del codice
- Creata la classe astratta Section per definire le caratteristiche comuni ad ogni classe dei Settings
