+++ src/main/java/moda/passwordmanager/frontend/GUIState.java
@@ -3,6 +3,5 @@ package moda.passwordmanager.frontend;
 public enum GUIState {
 
     ADD_DATA,
-    SHOW_DATA,
-    SETTINGS
+    SHOW_DATA
 }
